refactor-hooks-part-deux * Delete unused types * Use line defined in use-unique-lines
