+++ src/containers/Admin/EditTab/RealtimeDataPanel/PermanentLinesPanel.tsx
@@ -2,10 +2,11 @@ import React from 'react'
 import { ExpandablePanel } from '@entur/expand'
 import { ClosedLockIcon } from '@entur/icons'
 import { FilterChip } from '@entur/chip'
-import { DrawableRoute, Line } from '../../../../types'
+import { DrawableRoute } from '../../../../types'
 import { transportModeNameMapper } from '../../../../utils/utils'
 import { filterMap } from '../../../../utils/array'
 import { Settings } from '../../../../settings/settings'
+import { Line } from '../../../../logic/use-unique-lines/line'
 import { TransportModeIcon } from '../../../../components/TransportModeIcon/TransportModeIcon'
 import './linesPanel.scss'
