+++ src/containers/Admin/EditTab/RealtimeDataPanel/RealTimeDataPanel.tsx
@@ -4,8 +4,8 @@ import { Switch, TravelSwitch } from '@entur/form'
 import { Loader } from '@entur/loader'
 import { Heading3, Label, Paragraph } from '@entur/typography'
 import { ExpandablePanel } from '@entur/expand'
-import { Line } from '../../../../types'
 import { transportModeNameMapper } from '../../../../utils/utils'
+import { Line } from '../../../../logic/use-unique-lines/line'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { isTransport } from '../../../../utils/typeguards'
 import { TransportModeIcon } from '../../../../components/TransportModeIcon/TransportModeIcon'
